requirements.txt
absl-py==0.15.0
astunparse==1.6.3
beautifulsoup4==4.8.1
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2019.9.11
charset-normalizer==2.0.12
clang==5.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
gunicorn==20.0.4
h5py==3.1.0
idna==3.4
importlib-metadata==4.8.3
itsdangerous==1.1.0
Jinja2==2.10.3
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.7
MarkupSafe==1.1.1
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.1.2.30
opt-einsum==3.3.0
Pillow==6.2.1
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.9.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==3.141.0
six==1.15.0
soupsieve==1.9.5
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
tensorflow-hub==0.13.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.15
Werkzeug==0.16.0
wrapt==1.12.1
zipp==3.6.0
